--- HW/HW2 Docker/flask/app.py
@@ -1,19 +1,0 @@
-from flask import Flask, make_response
-
-app = Flask(__name__)
-
-@app.route('/')
-def hello():
-    response = make_response(
-        {
-            'response': 'Hello, World!',
-            'status': 200
-        }
-    )
-    return response
-
-if __name__ == '__main__':
-    # By default flask is only accessible from localhost.
-    # Set this to '0.0.0.0' to make it accessible from any IP address
-    # on your network (not recommended for production use)
-    app.run(host='0.0.0.0', debug=True)
