+++ HW/HW2 Docker/flask/container_storage/Dockerfile
@@ -2,7 +2,7 @@
 FROM ubuntu:22.04
 
 # Set the working directory inside the container
-WORKDIR /app
+WORKDIR /container_storage/app
 
 # Update and install system dependencies
 RUN apt-get update && apt-get install -y \
@@ -10,7 +10,7 @@ RUN apt-get update && apt-get install -y \
     python3-pip
 
 # Copy the current directory contents into the container at /app
-COPY . /app
+COPY . /container_storage/app
 
 # Normally we would install any needed packages specified in requirements.txt
 # RUN pip install --no-cache-dir -r requirements.txt
@@ -20,11 +20,12 @@ RUN pip install flask
 # Make port 5000 available to the world outside this container
 # Note that this is for the purposes of documentation
 # The port is actually exposed when you run the container from the command line
-EXPOSE 5000
+ENV PORT=5002
+EXPOSE $PORT
 
 # Run app.py when the container launches
-CMD ["python3", "app.py"]
+CMD ["python3", "container_storage/app/app.py"]
 
 # Health check (checks every 30 seconds to see if the endpoint returns a successful response)
-# HEALTHCHECK --interval=30s --timeout=3s \
-#  CMD curl -f http://localhost:5000/healthcheck || exit 1
+HEALTHCHECK --interval=30s --timeout=3s \
+CMD curl -f http://localhost:${PORT}/healthcheck || exit 1
